Create main.py
Have used the basic knowledge of pandas to build this small educational naming game, here we test our knowledge of the number of US states we know correctly.

extra feature:
we get to know what all states we've missed to that we can learn from the list. this is saved in a seperate csv file, triggered when we enter "Exit" in the answer input box.